requirements.txt
locust>=2.10.0
msal>=1.26.0
requests>=2.31.0
websocket-client>=1.7.0
cryptography>=42.0.0
keyring>=24.0.0
python-dotenv>=1.0.0
rich>=13.0.0
colorama>=0.4.6
